backend python/requirements.txt
fastapi>=0.110
uvicorn>=0.27

python-dotenv>=1.0
sqlalchemy>=2.0

pydantic>=2.6
pydantic-settings>=2.2

langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.7

faiss-cpu>=1.8.0

requests>=2.31
beautifulsoup4>=4.12

pytest>=8.0
pytest-asyncio>=0.23
